Fix for Linux.
This fix allows the touchscreen to work on reboot to Linux. It does not affect Windows power management in hibernation, sleep, or Fast Shutdown. The fix is only triggered when Exit is to WdfPowerDeviceD3Final, which is State 5 seen here:

https://msdn.microsoft.com/en-us/library/windows/hardware/ff552421(v=vs.85).aspx

FxPreviousState is also renamed to FxTargetState in accordance with the descriptions here:


https://msdn.microsoft.com/en-us/library/windows/hardware/ff556803(v=vs.85).aspx

https://github.com/Microsoft/Windows-driver-samples/blob/master/sensors/SimpleDeviceOrientationSensor/device.cpp#L537

This change is neutral to Windows functionality, but provides an alternative to cold booting to get the touchscreen to work in Linux